N-API mapnik.VectorTile clear/render (WIP) [skip ci]
